TWELVE STEP: 1.Order and OrderItem entities, controllers, services, repositories, mappers and dtos implemented 2.Database tables created by liquibase 3. Data validation developed 4. Code style improved

+++ src/main/java/app/service/CategoryServiceImpl.java
@@ -51,6 +51,10 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void deleteById(Long id) {
+        if (categoryRepository.findById(id).isEmpty()) {
+            throw new EntityNotFoundException("Category with id: " + id + " not found");
+        }
+
         categoryRepository.deleteById(id);
     }
 }
